image processing integrated

+++ android-studio/BabyBox/app/src/main/java/com/babybox/fragment/AbstractFeedViewFragment.java
@@ -197,7 +197,7 @@ public abstract class AbstractFeedViewFragment extends TrackedFragment {
     protected void attachEndlessScrollListener() {
         feedView.setOnScrollListener(new EndlessScrollListener(layoutManager) {
             @Override
-            public void onLoadMore(int page) {
+            public void onLoadMore(Long page) {
                 loadFeed(Long.valueOf(page - 1), getFeedFilter());
             }
         });
